+++ src/StarknetProvider/index.tsx
@@ -46,3 +46,41 @@ export function StarknetProvider({ children }: { children: React.ReactNode }) {
     </StarknetConfig>
   );
 }
+
+// "use client"
+// import React from "react";
+
+// import { sepolia } from "@starknet-react/chains";
+// import {
+//   StarknetConfig,
+//   publicProvider,
+//   argent,
+//   braavos,
+//   useInjectedConnectors,
+//   voyager,
+//   InjectedConnector
+// } from "@starknet-react/core";
+// import { ArgentMobileConnector } from "starknetkit/argentMobile";
+// import { WebWalletConnector } from "starknetkit/webwallet";
+
+// export function StarknetProvider({ children }: { children: React.ReactNode }) {
+//   const connectors = [
+//     new InjectedConnector({ options: {id: "braavos", name: "Braavos" }}),
+//     new InjectedConnector({ options: {id: "argentX", name: "Argent X" }}),
+//     new ArgentMobileConnector(),
+//   ]
+
+//   const chains = [sepolia]
+
+//   return (
+//     <StarknetConfig
+//       chains={chains}
+//       provider={publicProvider()}
+//       connectors={connectors}
+//       explorer={voyager}
+//       autoConnect
+//     >
+//       {children}
+//     </StarknetConfig>
+//   );
+// }
